Creates separate component for table
