+++ simple-booking-webapp/components/HowItWorks/index.tsx
@@ -10,7 +10,7 @@ const HowItWorks: FC = () => {
   return (
     <div className='grid grid-cols-1 md:grid-cols-3 gap-4 mb-8'>
       <Panel
-        header='Fill up the form with your details'
+        header='Fill up our booking form with your details'
         description='No sign-up necessary. We just ask for a name, some comments and a password.'
         icon={<EditNote width={ICON_SIZE} height={ICON_SIZE} />}
       />
