+++ simple-booking-webapp/utils/data.ts
@@ -50,49 +50,49 @@ export const timeSlots = [
 export const bookingTableData: Appointment[] = [
   {
     id: '1',
-    date: moment('October 23, 2023 08:00:00').toDate(),
+    date: moment('October 24, 2023 08:00:00').toDate(),
     name: 'Tony Reichert',
     comments: 'This is my comment about my booking information',
     password: '1234',
   },
   {
     id: '2',
-    date: moment('October 24, 2023 09:00:00').toDate(),
+    date: moment('October 25, 2023 09:00:00').toDate(),
     name: 'Zoey Lang',
     comments: 'This is my comment about my booking information',
     password: 'abcd',
   },
   {
     id: '3',
-    date: moment('October 25, 2023 10:00:00').toDate(),
+    date: moment('October 26, 2023 10:00:00').toDate(),
     name: 'Jane Fisher',
     comments: 'This is my comment about my booking information',
     password: '1234',
   },
   {
     id: '4',
-    date: moment('October 26, 2023 13:00:00').toDate(),
+    date: moment('October 27, 2023 13:00:00').toDate(),
     name: 'William Howard',
     comments: 'This is my comment about my booking information',
     password: 'abcd',
   },
   {
     id: '5',
-    date: moment('October 26, 2023 15:00:00').toDate(),
+    date: moment('October 27, 2023 15:00:00').toDate(),
     name: 'Kristen Copper',
     comments: 'This is my comment about my booking information',
     password: '1234',
   },
   {
     id: '6',
-    date: moment('October 23, 2023 09:00:00').toDate(),
+    date: moment('October 24, 2023 09:00:00').toDate(),
     name: 'Tony Reichert II',
     comments: 'This is my comment about my booking information',
     password: '1234',
   },
   {
     id: '7',
-    date: moment('October 23, 2023 11:00:00').toDate(),
+    date: moment('October 24, 2023 11:00:00').toDate(),
     name: 'Tony Reichert III',
     comments: 'This is my comment about my booking information',
     password: '1234',
